Rename frontend game files to use battle
