+++ packages/client/src/components/BattleControlButtons.tsx
