+++ packages/client/src/components/NoActionsDialog.tsx
@@ -7,7 +7,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog';
-import { useGame } from '@/contexts/GameContext';
+import { useGame } from '@/contexts/BattleContext';
 
 import { Button } from './ui/button';
