+++ packages/client/src/components/NoBattleScreen.tsx
