+++ packages/client/src/components/TowerSelection.tsx
@@ -1,8 +1,8 @@
 import { useDndContext } from '@dnd-kit/core';
 
+import { INSTALLABLE_TOWERS } from '@/components/BattleBoard';
 import { Draggable } from '@/components/Draggable';
-import { INSTALLABLE_TOWERS } from '@/components/GameBoard';
-import { useGame } from '@/contexts/GameContext';
+import { useGame } from '@/contexts/BattleContext';
 
 export const TowerSelection = (): JSX.Element => {
   const { activeTowerId, handleTowerSelect, isPlayer1 } = useGame();
