+++ packages/client/src/pages/Battle.tsx
@@ -19,19 +19,19 @@ import { useAccount } from 'wagmi';
 
 import { BackgroundAnimation } from '@/components/BackgroundAnimation';
 import { BatteryInfoDialog } from '@/components/BatteryInfoDialog';
+import { GameBoard, INSTALLABLE_TOWERS } from '@/components/BattleBoard';
+import { GameControlButtons } from '@/components/BattleControlButtons';
+import { GameStatusBar } from '@/components/BattleStatusBar';
 import { CastleHitDialog } from '@/components/CastleHitDialog';
 import { ForfeitDialog } from '@/components/ForfeitDialog';
-import { GameBoard, INSTALLABLE_TOWERS } from '@/components/GameBoard';
-import { GameControlButtons } from '@/components/GameControlButtons';
-import { GameStatusBar } from '@/components/GameStatusBar';
 import { HowToPlayDialog } from '@/components/HowToPlayDialog';
 import { LoadingScreen } from '@/components/LoadingScreen';
 import { NoActionsDialog } from '@/components/NoActionsDialog';
-import { NoGameScreen } from '@/components/NoGameScreen';
+import { NoGameScreen } from '@/components/NoBattleScreen';
 import { PlayAgainDialog } from '@/components/PlayAgainDialog';
 import { TowerSelection } from '@/components/TowerSelection';
 import { Button } from '@/components/ui/button';
-import { GameProvider, useGame } from '@/contexts/GameContext';
+import { GameProvider, useGame } from '@/contexts/BattleContext';
 import { useSettings } from '@/contexts/SettingsContext';
 import { useMUD } from '@/hooks/useMUD';
 import { BATTERY_STORAGE_LIMIT } from '@/utils/constants';
